pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kresearch"
version = "0.1.0"
description = "Autonomous deep research agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "ddgs>=7.0",
    "trafilatura>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5", "mypy>=1.10"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]

[project.scripts]
kresearch = "kresearch.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/kresearch"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
